As per review comments, optimize imports and small refactorings
