+++ plugin-infra/go-plugin-access/src/main/java/com/thoughtworks/go/plugin/access/pluggabletask/TaskExtension.java
@@ -18,7 +18,6 @@ package com.thoughtworks.go.plugin.access.pluggabletask;
 
 import com.thoughtworks.go.plugin.access.PluginRequestHelper;
 import com.thoughtworks.go.plugin.access.common.AbstractExtension;
-import com.thoughtworks.go.plugin.access.common.settings.MessageHandlerForPluginSettingsRequestProcessor;
 import com.thoughtworks.go.plugin.access.common.settings.PluginSettingsJsonMessageHandler1_0;
 import com.thoughtworks.go.plugin.api.response.execution.ExecutionResult;
 import com.thoughtworks.go.plugin.api.response.validation.ValidationResult;
@@ -33,6 +32,7 @@ import org.springframework.stereotype.Component;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 import static com.thoughtworks.go.plugin.domain.common.PluginConstants.PLUGGABLE_TASK_EXTENSION;
 
@@ -70,7 +70,7 @@ public class TaskExtension extends AbstractExtension {
     }
 
     @Override
-    protected MessageHandlerForPluginSettingsRequestProcessor messageHandlerForPluginSettingsRequestProcessor(String pluginVersion) {
+    public String pluginSettingsJSON(String pluginId, Map<String, String> pluginSettings) {
         throw new UnsupportedOperationException("Fetch PluginSettings is not supported by Task Endpoint.");
     }
